refactor(ui): consolidate connection ownership and split control shell
