+++ ui/src/lib/sessions/navigation-handoff.ts
@@ -1,17 +1,30 @@
 type SessionNavigationHandoff = {
   pathname: string;
   sessionKey: string;
+  client: object | null;
+  hello: object | null;
+};
+
+type SessionNavigationHandoffOwner = {
+  readonly snapshot: {
+    readonly client: object | null;
+    readonly hello: object | null;
+  };
 };
 
 const SESSION_NAVIGATION_HANDOFF_TTL_MS = 2_000;
-const sessionNavigationHandoffs = new WeakMap<object, SessionNavigationHandoff>();
+const sessionNavigationHandoffs = new WeakMap<
+  SessionNavigationHandoffOwner,
+  SessionNavigationHandoff
+>();
 
 export function prepareSessionNavigationHandoff(
-  owner: object,
+  owner: SessionNavigationHandoffOwner,
   pathname: string,
   sessionKey: string,
 ): void {
-  const handoff = { pathname, sessionKey };
+  const { client, hello } = owner.snapshot;
+  const handoff = { pathname, sessionKey, client, hello };
   sessionNavigationHandoffs.set(owner, handoff);
   globalThis.setTimeout(() => {
     if (sessionNavigationHandoffs.get(owner) === handoff) {
@@ -21,7 +34,7 @@ export function prepareSessionNavigationHandoff(
 }
 
 export function consumeSessionNavigationHandoff(
-  owner: object,
+  owner: SessionNavigationHandoffOwner,
   pathname: string,
 ): string | undefined {
   const handoff = sessionNavigationHandoffs.get(owner);
@@ -29,5 +42,11 @@ export function consumeSessionNavigationHandoff(
     return undefined;
   }
   sessionNavigationHandoffs.delete(owner);
+  // Browser clients survive transport reconnects, so hello identity must also
+  // match before an old connection can bypass authoritative route resolution.
+  const { client, hello } = owner.snapshot;
+  if (handoff.client !== client || handoff.hello !== hello) {
+    return undefined;
+  }
   return handoff.sessionKey;
 }
