+++ ui/src/lib/sessions/session-roster-refresh.ts
@@ -1,0 +1,315 @@
+import type { SessionsListResult } from "../../api/types.ts";
+import { createSessionEventRefreshCoordinator } from "./event-refresh-coordinator.ts";
+import type {
+  SessionConnectionOwner,
+  SessionGateway,
+  SessionListOptions,
+  SessionRefreshOptions,
+  SessionState,
+} from "./session-capability.ts";
+import {
+  areUiSessionKeysEquivalent,
+  normalizeAgentId,
+  parseAgentSessionKey,
+  resolveUiSelectedGlobalAgentId,
+  uiSessionRowMatchesSelectedChat,
+} from "./session-key.ts";
+import { DEFAULT_SESSION_LIST_QUERY, requestSessionList } from "./session-requests.ts";
+
+type SessionRosterRefreshHost = {
+  connection: SessionConnectionOwner;
+  snapshot: () => SessionGateway["snapshot"];
+  readState: () => SessionState;
+  publish: (state: SessionState, errorSource?: "session-observer" | "operation") => void;
+  observerError: () => string | null;
+  decorate: (result: SessionsListResult | null) => SessionsListResult | null;
+  onCanonicalList: (result: SessionsListResult | null) => void;
+};
+
+function appendSessionResults(
+  previous: SessionsListResult,
+  page: SessionsListResult,
+): SessionsListResult {
+  const seen = new Set<string>();
+  const sessions = [...previous.sessions, ...page.sessions].filter((row) => {
+    if (!row.key || seen.has(row.key)) {
+      return false;
+    }
+    seen.add(row.key);
+    return true;
+  });
+  const totalCount = page.totalCount ?? previous.totalCount;
+  const hasMore =
+    page.hasMore ??
+    (typeof totalCount === "number" && Number.isFinite(totalCount)
+      ? sessions.length < totalCount
+      : false);
+  return {
+    ...page,
+    count: sessions.length,
+    totalCount,
+    hasMore,
+    nextOffset: page.nextOffset ?? (hasMore ? sessions.length : null),
+    sessions,
+  };
+}
+
+export function createSessionRosterRefresh(host: SessionRosterRefreshHost) {
+  let inFlight: Promise<void> | null = null;
+  let queuedExplicitRefresh: SessionRefreshOptions | null = null;
+  let eventRefreshQueued = false;
+  let lastListOptions: SessionListOptions = {};
+  let hasForegroundListOptions = false;
+  let hasSeededListOptions = false;
+
+  const list = async (options: SessionListOptions = {}): Promise<SessionsListResult | null> => {
+    const scope = host.connection.capture();
+    if (!scope) {
+      return null;
+    }
+    const result = await requestSessionList(scope.client, options);
+    return host.connection.isCurrent(scope) ? host.decorate(result ?? null) : null;
+  };
+
+  const load = async (options: SessionRefreshOptions) => {
+    const scope = host.connection.capture();
+    if (!scope) {
+      return;
+    }
+    const { append = false, force: _force, backgroundHydrate = false, ...requestOptions } = options;
+    const durableListOptions: SessionListOptions = { ...requestOptions };
+    // Pagination is request-local; replacements retain filters but restart at page one.
+    delete durableListOptions.offset;
+    if (!backgroundHydrate) {
+      lastListOptions = durableListOptions;
+      hasForegroundListOptions = true;
+    } else if (!hasForegroundListOptions && !hasSeededListOptions) {
+      lastListOptions = durableListOptions;
+      hasSeededListOptions = true;
+    }
+    if (!backgroundHydrate) {
+      const error = host.observerError();
+      host.publish(
+        { ...host.readState(), loading: true, error, deletedSessions: [] },
+        error ? "session-observer" : undefined,
+      );
+    }
+    try {
+      const result = await requestSessionList(scope.client, requestOptions);
+      if (!host.connection.isCurrent(scope)) {
+        return;
+      }
+      const currentState = host.readState();
+      let nextResult =
+        result && append && requestOptions.offset && currentState.result
+          ? appendSessionResults(currentState.result, result)
+          : result;
+      if (append && nextResult && !backgroundHydrate) {
+        // Canonical event refreshes must retain all previously appended visible pages.
+        lastListOptions = {
+          ...durableListOptions,
+          limit: Math.max(
+            durableListOptions.limit ?? DEFAULT_SESSION_LIST_QUERY.limit,
+            nextResult.sessions.length,
+          ),
+        };
+      }
+      if (backgroundHydrate && nextResult) {
+        const snapshot = host.snapshot();
+        const currentKey = snapshot.sessionKey?.trim();
+        if (currentKey) {
+          const currentAgentId = normalizeAgentId(
+            parseAgentSessionKey(currentKey)?.agentId ?? resolveUiSelectedGlobalAgentId(snapshot),
+          );
+          const previousCurrentRow =
+            currentState.result?.sessions.find((row) =>
+              areUiSessionKeysEquivalent(row.key, currentKey),
+            ) ??
+            (currentState.agentId === currentAgentId
+              ? currentState.result?.sessions.find((row) =>
+                  uiSessionRowMatchesSelectedChat(snapshot, row.key, currentKey),
+                )
+              : undefined);
+          if (
+            previousCurrentRow &&
+            !nextResult.sessions.some((row) =>
+              uiSessionRowMatchesSelectedChat(snapshot, row.key, currentKey),
+            )
+          ) {
+            const sessions = [...nextResult.sessions, previousCurrentRow];
+            nextResult = { ...nextResult, count: sessions.length, sessions };
+          }
+        }
+      }
+      nextResult = host.decorate(nextResult);
+      host.onCanonicalList(nextResult);
+      const state = host.readState();
+      const error = host.observerError();
+      host.publish(
+        {
+          result: nextResult,
+          agentId: requestOptions.agentId?.trim() ? normalizeAgentId(requestOptions.agentId) : null,
+          modelOverrides: state.modelOverrides,
+          loading: backgroundHydrate ? state.loading : false,
+          error,
+          deletedSessions: [],
+          groups: state.groups,
+          sectionOrder: state.sectionOrder,
+        },
+        error ? "session-observer" : undefined,
+      );
+    } catch (error) {
+      if (host.connection.isCurrent(scope)) {
+        const state = host.readState();
+        host.publish(
+          {
+            ...state,
+            loading: backgroundHydrate ? state.loading : false,
+            error: String(error),
+            deletedSessions: [],
+          },
+          "operation",
+        );
+      }
+    }
+  };
+
+  const absorbPendingEventRefresh = () => {
+    eventRefreshCoordinator.absorb();
+    eventRefreshQueued = false;
+  };
+
+  const takeNextQueuedRefresh = (): SessionRefreshOptions | null => {
+    const explicitRefresh = queuedExplicitRefresh;
+    queuedExplicitRefresh = null;
+    if (explicitRefresh) {
+      // Replacement absorbs earlier events; append still needs its trailing replacement.
+      if (explicitRefresh.append !== true) {
+        absorbPendingEventRefresh();
+      }
+      return explicitRefresh;
+    }
+    if (!eventRefreshQueued) {
+      return null;
+    }
+    eventRefreshQueued = false;
+    return { ...lastListOptions, force: true };
+  };
+
+  const drainRefreshQueue = async (options: SessionRefreshOptions) => {
+    const scope = host.connection.capture();
+    if (!scope) {
+      return;
+    }
+    let next: SessionRefreshOptions | null = options;
+    while (next) {
+      await load(next);
+      if (!host.connection.isCurrent(scope)) {
+        return;
+      }
+      next = takeNextQueuedRefresh();
+    }
+  };
+
+  const startRefresh = (options: SessionRefreshOptions) => {
+    const request = drainRefreshQueue(options).finally(() => {
+      if (inFlight === request) {
+        inFlight = null;
+      }
+    });
+    inFlight = request;
+    return request;
+  };
+
+  const refresh = (options: SessionRefreshOptions = {}): Promise<void> => {
+    if (!host.connection.capture()) {
+      return Promise.resolve();
+    }
+    if (inFlight) {
+      queuedExplicitRefresh = options;
+      return inFlight;
+    }
+    const hasListOverrides = Object.entries(options).some(
+      ([key, value]) => key !== "force" && key !== "backgroundHydrate" && value !== undefined,
+    );
+    if (host.readState().result && !options.force && !hasListOverrides) {
+      return Promise.resolve();
+    }
+    if (options.append !== true) {
+      absorbPendingEventRefresh();
+    }
+    return startRefresh(options);
+  };
+
+  const refreshFromEvent = () => {
+    if (!host.connection.capture()) {
+      return Promise.resolve();
+    }
+    if (inFlight) {
+      eventRefreshQueued = true;
+      return inFlight;
+    }
+    return startRefresh({ ...lastListOptions, force: true });
+  };
+
+  const eventRefreshCoordinator = createSessionEventRefreshCoordinator({
+    canRefresh: () => host.connection.capture() !== null,
+    refresh: refreshFromEvent,
+  });
+  const flushEventRefresh = () => eventRefreshCoordinator.flush();
+
+  const handleVisibilityChange = () => {
+    if (document.visibilityState === "hidden") {
+      flushEventRefresh();
+    }
+  };
+  const observesPageLifecycle =
+    typeof document !== "undefined" && typeof globalThis.addEventListener === "function";
+  if (observesPageLifecycle) {
+    document.addEventListener("visibilitychange", handleVisibilityChange);
+    globalThis.addEventListener("pagehide", flushEventRefresh);
+  }
+
+  const refreshReplacement = (agentId?: string | null): Promise<void> => {
+    const options = {
+      ...lastListOptions,
+      includeDerivedTitles: lastListOptions.includeDerivedTitles ?? true,
+    };
+    const normalizedAgentId = agentId?.trim();
+    if (normalizedAgentId) {
+      options.agentId = normalizedAgentId;
+    }
+    return refresh({ ...options, force: true });
+  };
+
+  return {
+    list,
+    refresh,
+    refreshReplacement,
+    setCreatorFilter(creatorId: string | null) {
+      const options = { ...lastListOptions, creatorId: creatorId?.trim() || undefined };
+      delete options.offset;
+      return refresh({ ...options, force: true });
+    },
+    lastOptions: () => lastListOptions,
+    scheduleEvent: () => eventRefreshCoordinator.schedule(),
+    reset() {
+      eventRefreshCoordinator.reset();
+      inFlight = null;
+      queuedExplicitRefresh = null;
+      eventRefreshQueued = false;
+    },
+    dispose() {
+      // Flush before disposal so page-exit events start the trailing canonical list.
+      flushEventRefresh();
+      eventRefreshCoordinator.dispose();
+      if (observesPageLifecycle) {
+        document.removeEventListener("visibilitychange", handleVisibilityChange);
+        globalThis.removeEventListener("pagehide", flushEventRefresh);
+      }
+      inFlight = null;
+      queuedExplicitRefresh = null;
+      eventRefreshQueued = false;
+    },
+  };
+}
