+++ ui/src/pages/chat/chat-pane-context.ts
@@ -7,6 +7,7 @@ import {
 } from "../../app/question-prompt.ts";
 import { loadSettings } from "../../app/settings.ts";
 import { readPresenceEntries } from "../../app/user-profile.ts";
+import { createGatewayConnectionLifecycle } from "../../lib/gateway-connection-lifecycle.ts";
 import { isGatewayMethodAdvertised } from "../../lib/gateway-methods.ts";
 import { parseCatalogSessionKey } from "../../lib/sessions/catalog-key.ts";
 import { resolveSessionKey } from "../../lib/sessions/index.ts";
@@ -41,6 +42,14 @@ import { normalizeSidebarLayout } from "./sidebar-layout.ts";
 import { reconcileWaitingApprovalsFromSnapshot } from "./tool-stream.ts";
 
 export abstract class ChatPaneContext extends ChatPaneLifecycle {
+  private gatewayConnectionLifecycle?: ReturnType<typeof createGatewayConnectionLifecycle>;
+
+  override disconnectedCallback() {
+    this.gatewayConnectionLifecycle?.dispose();
+    this.gatewayConnectionLifecycle = undefined;
+    super.disconnectedCallback();
+  }
+
   protected applySessionsState(stateValue: ApplicationContext["sessions"]["state"]) {
     const state = this.state;
     if (!state) {
@@ -160,8 +169,12 @@ export abstract class ChatPaneContext extends ChatPaneLifecycle {
     const previousMediaAuthToken = resolveAssistantAttachmentAuthToken(state);
     const wasConnected = state.connected;
     const previousSidebarSessionKey = canonicalUiSessionKeyForPersistence(state, state.sessionKey);
-    const sourceChanged =
-      state.client !== snapshot.client || wasConnected !== (snapshot.phase === "connected");
+    const connectionLifecycle = (this.gatewayConnectionLifecycle ??=
+      createGatewayConnectionLifecycle({
+        client: state.client,
+        phase: state.connected ? "connected" : "stopped",
+      }));
+    const sourceChanged = connectionLifecycle.transition(snapshot);
     const clientChanged = this.connectedClient !== snapshot.client;
     if (snapshot.phase !== "connected") {
       this.presencePayload = undefined;
